# oas/2.0/officers.yaml
openapi: "3.0.3"
info:
  title: "Officers"
  description: "API Description"
  version: "0.7.0"
servers:
  - url: "http://127.0.0.1:5001/api/v1"
    description: "Development environment"
  - url: "https://dev.nationalpolicedata.org/api/v1"
    description: "Staging environment"
  - url: "https://api.nationalpolicedata.org"
    description: "Production environment"
...
security:
  - bearerAuth: []
...
paths:
  /officers:
    get:
      tags:
        - "Officers"
      summary: "Get officers"
      operationId: "getOfficers"
      description: >
        Returns a list of officer. A number of simple filters can be applied to limit the results. If no filters
        are applied, a random sample of officers will be returned.
      parameters: 
        - $ref: '../common/pagination.yaml#/components/parameters/page'
        - $ref: '../common/pagination.yaml#/components/parameters/per_page'
        - $ref: '#/components/parameters/query'
        - $ref: '#/components/parameters/first_name'
        - $ref: '#/components/parameters/middle_name'
        - $ref: '#/components/parameters/last_name'
        - $ref: '#/components/parameters/suffix'
        - $ref: '#/components/parameters/active_after'
        - $ref: '#/components/parameters/active_before'
        - $ref: '#/components/parameters/agency'
        - $ref: '#/components/parameters/rank'
        - $ref: '#/components/parameters/unit'
        - $ref: '#/components/parameters/ethnicity'
        - $ref: '#/components/parameters/badge_number'
        - $ref: '#/components/parameters/search_result'
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficerList"
        '401':
          $ref: '../common/error.yaml#/components/responses/unauthorizedError'
    post:
      tags: 
        -  Officers
      summary: "Create a new officer"
      operationId: "createOfficer"
      requestBody: 
        content: 
          application/json: 
            schema: 
              $ref: "#/components/schemas/CreateOfficer"
      responses:
        '201':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Officer'
        '400':
          $ref: '../common/error.yaml#/components/responses/validationError'
        '401':
          $ref: '../common/error.yaml#/components/responses/unauthorizedError'
...
  /officers/{uid}:
    parameters: 
      - name: uid
        in: path
        description: UID of the officer
        required: true
        schema:
          type: string
    get:
      tags:
        - "Officers"
      summary: "Get officer by uid"
      operationId: "getOfficerById"
      parameters:
        - $ref: '#/components/parameters/include_officer'
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
        '404':
          $ref: '../common/error.yaml#/components/responses/notFoundError'
        '401':
          $ref: '../common/error.yaml#/components/responses/unauthorizedError'
    patch:
      tags:
        - "Officers"
      summary: "Update an existing officer"
      operationId: "updateOfficer"
      requestBody:
        description: "Officer object that needs to be updated"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Officer"
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
        '400':
          $ref: '../common/error.yaml#/components/responses/validationError'
        '401':
          $ref: '../common/error.yaml#/components/responses/unauthorizedError'
    delete:
      tags:
        - "Officers"
      summary: "Delete an officer"
      operationId: "deleteOfficer"
      description: "Delete an officer by uid"
      parameters:
        - name: "uid"
          in: "path"
          description: "UID of officer to delete"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "204":
          description: "No content"
        '404':
          $ref: '../common/error.yaml#/components/responses/notFoundError'
        '401':
          $ref: '../common/error.yaml#/components/responses/unauthorizedError'
  /officers/{uid}/employment:
    parameters: 
      - name: uid
        in: path
        description: UID of the officer
        required: true
        schema:
          type: string
    get:
      summary: "Get Employment History"
      operationId: "getEmploymentHistory"
      description: >
        Get employment history for an officer. This includes all the agencies
        the officer has worked for.
      tags:
        - "Employment History"
      responses: 
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmploymentList'
        '401':
          $ref: '../common/error.yaml#/components/responses/unauthorizedError'
    put:
      summary: "Update Employment History"
      operationId: "updateEmploymentHistory"
      description: >
        Update the employment history for an officer. This includes all the agencies
        the officer has worked for.
      tags:
        - "Employment History"
      requestBody:
        content: 
          application/json: 
            schema: 
              $ref: '#/components/schemas/AddEmploymentList'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/EmploymentList'
        '400':
          $ref: '../common/error.yaml#/components/responses/validationError'
        '401':
          $ref: '../common/error.yaml#/components/responses/unauthorizedError'
...
components:
  parameters: 
    query:
      name: query
      in: query
      description: >
        A general search query to filter officers with. This will search across
        first, middle, and last names.
      required: false
      schema:
        type: string
    active_after:
      name: active_after
      in: query
      description: >
        Filter officers who were active after this date. The date should be in
        the format YYYY-MM-DD.
      required: false
      schema:
        type: string
        format: date
    active_before:
      name: active_before
      in: query
      description: >
        Filter officers who were active before this date. The date should be in the format YYYY-MM-DD.
      required: false
      schema:
        type: string
        format: date
    agency:
      name: agency
      in: query
      description: >
        Filter officers by the agency they are employed by. The value should be a comma-separated list of agency uids.
      required: false
      schema:
        type: string
    rank:
      name: rank
      in: query
      description: >
        Filter officers by their rank. The value should be a comma-separated list of ranks.
      required: false
      schema:
        type: string
    unit:
      name: unit
      in: query
      description: >
        Filter officers by their unit. The value should be a comma-separated list of unit uids.
      required: false
      schema:
        type: string
    first_name:
      name: first_name
      in: query
      description: >
        Filter officers by their first name. The value should be a string.
      required: false
      schema:
        type: string
    middle_name:
      name: middle_name
      in: query
      description: >
        Filter officers by their middle name. The value should be a string.
      required: false
      schema:
        type: string
    last_name:
      name: last_name
      in: query
      description: >
        Filter officers by their last name. The value should be a string.
      required: false
      schema:
        type: string
    suffix:
      name: suffix
      in: query
      description: >
        Filter officers by their namesuffix. The value should be a string.
      required: false
      schema:
        type: string
    ethnicity:
      name: ethnicity
      in: query
      description: >
        Filter officers by their ethnicity. The value should be a comma-separated list of ethnicities.
      required: false
      schema:
        type: string
    badge_number:
      name: badge_number
      in: query
      description: >
        Filter officers by their badge number. The value should be a comma-separated list of badge numbers.
      required: false
      schema:
        type: string
    search_result:
      name: search_result
      in: query
      description: >
        Set as true to return officers in a search result format, which includes highlights
        and relevance scores.
      required: false
      schema:
        type: boolean
...
    include_officer:
      name: include
      in: query
      description: >
        Request additional officer details to be included in the metrics
        response. You can select from the following options:

          - `employment`: Include the officer's recent employment history.
          - `allegations`: Include a summary the allegations made against the officer.
      required: false
      schema:
        type: string
        enum:
          - employment
          - allegations
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AddEmployment:
      allOf: 
        - $ref: "#/components/schemas/BaseEmployment"
        - type: "object"
        - required:
          - unit_uid
          - badge_number
...
    AddEmploymentList:
      type: object
      properties:
        agencies:
          type: "array"
          description: "The units to add to the officer's employment history."
          items:
            $ref: "#/components/schemas/AddEmployment"
    Employment:
      allOf: 
        - $ref: "#/components/schemas/BaseEmployment"
...
    EmploymentList:
      allOf:
        - $ref: '../common/pagination.yaml#/components/schemas/PaginatedResponse'
        - type: "object"
          properties:
            results:
              type: "array"
              items:
                $ref: "#/components/schemas/Employment"
    BaseOfficer:
      type: "object"
      properties:
        first_name:
          type: "string"
          description: "First name of the officer"
        middle_name:
          type: "string"
          description: "Middle name of the officer"
        last_name:
          type: "string"
          description: "Last name of the officer"
        suffix:
          type: "string"
          description: "Suffix of the officer's name"
        ethnicity:
          type: "string"
          description: "The ethnicity of the officer"
          enum:
            - American Indian or Alaska Native
            - Asian
            - Black or African American
            - Hispanic or Latino
            - Native Hawaiian or Other Pacific Islander
            - White
        gender:
          type: "string"
          description: "The gender of the officer"
          enum:
            - Male
            - Female
            - Other
        date_of_birth:
          type: "string"
          format: "date"
          description: "The date of birth of the officer"
        state_ids:
          type: "array"
          description: "The state ids of the officer"
          items:
            $ref: "#/components/schemas/StateId"
    CreateOfficer:
      allOf:
        - $ref: "#/components/schemas/BaseOfficer"
        - type: "object"
        - required:
          - first_name
          - last_name
...
    Officer:
      allOf: 
        - $ref: "#/components/schemas/BaseOfficer"
        - type: "object"
        - properties:
            uid:
              type: "string"
              description: "The uid of the officer"
            employment_history:
              type: "string"
              description: "A link to retrieve the employment history of the officer"
            allegations:
              type: "string"
              description: "A link to retrieve the allegations against the officer"
            litigation:
              type: "string"
              description: "A link to retrieve the litigation against the officer"
    OfficerList:
      allOf: 
        - $ref: '../common/pagination.yaml#/components/schemas/PaginatedResponse'
        - type: "object"
          properties:
            results:
              type: "array"
              items:
                $ref: "#/components/schemas/Officer"
    StateId:
      type: "object"
      properties:
        uid:
          type: "string"
          description: "The UUID of this state id"
        state:
          type: "string"
          description: "The state of the state id"
        id_name:
          type: "string"
          description: "The name of the id. For example, Tax ID, Driver's License, etc."
        value:
          type: "string"
          description: "The value of the id."
...